Workaround for disabled SwipeRefresh breaking other Functionalities
If we just disabled the SwipeRefreshLayout entirely,
the RecyclerView inside this container did not work properly, anymore.
For example, when re-ordering Website entries,
the result of the re-order was not viewable
by the user.
If we just cancel the Refresh action,
when swipeRefreshIsEnabled == false,
then it still swipes,
but does not actually refresh the WebsiteStatus list.

This middle ground is acceptable, for now.