[close #22] Fix Search Filter not working reliably all the time
Whenever all Website Entries' status were refreshed,
there was a change in the WebsiteEntryList's LiveData,
which lead to the Observer triggering
and calling `webSiteEntryAdapter.setAllTodoItems(it)`.
Because of this, the search would not filter any entries, after all,
because all available Website Entries would've been loaded
by the Refresh, anyway, overwriting any filter
applied by the searchView.

To overcome this issue, `webSiteEntryAdapter.setAllTodoItems(it)`
will only be called, if the searchView is iconified,
i.e. minimised, i.e. not in use, i.e. the list ist not being filtered.