Improve Website Favicon Retrieval Reliability
https://github.com/bumptech/glide/issues/4704

On older Android API version, it may be the case, that
the fallback icon does not load, as the target website
appears to serve the icon with the help of a self-signed certificate.
Glide errors out in these cases, as described in the issue above.

To make icon display work on lower Android API versions, as well,
the icon gets loaded from the local image asset storage,
as a last resort.

This is the order of precedence for fetching a Website icon:
1. The actual icon corresponding to the website in question.
2. The placeholder icon retrieved from the fallback URL.
3. The placeholder icon loaded from local image asset storage.

This way, icons will be shown on as many Android versions, as supported.