[#35] Add Website Entry Tags & Add Discard Changes Confirmation
Refactored and improved a lot of code,
especially the code regarding Tag Cloud.

To avoid problems with memorised entry data,
the user is asked to explicitly discard the data,
when pressing any back button.
This solves issues with the wrong data being
memorised or data being memorised,
even though the user wants it discarded.